requirements.txt
argon2-cffi==18.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Django==2.1
django-anymail==3.0
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-environ==0.4.5
django-gravatar2==1.4.2
django-mathfilters==0.4.0
django-zxcvbn-password==2.0.1
idna==2.7
mysqlclient==1.3.13
packaging==17.1
pip-review==1.0
psycopg2>=2.7,<2.8 --no-binary psycopg2
pycparser==2.18
pyparsing==2.2.0
pytz==2018.5
requests==2.19.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
zxcvbn-python==4.4.24
